ignore order switch, Verification data
Added paramter to ignore the order of json-items when verifying the result. APIC sometimes returns multiple childs in some unpredictable ordered way.
Can be set via -m or by calling verify_ignore_order() in the task
Also added output of live-data when the verification fails to ease troubleshooting (diff is nice, but sometimes the big picture makes it clearer)